% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_preprocessing_functions.R
\name{block_scale}
\alias{block_scale}
\title{Scales a block of data to unit variance}
\usage{
block_scale(input_data, settings = list(block_std = NULL))
}
\arguments{
\item{input_data}{A matrix containing data. The rows indicate the samples,
the columns indicate the variables.}

\item{settings}{A list with a single named element "block_std". When
block_std is set, the supplied value is used for scaling, otherwise it is
calculated.}
}
\value{
A named list with "preprocessed_data": the scaled input_data matrix,
  and "settings", which is equal to the settings input parameter when set,
  and otherwise contains the calculated and used block_std.
}
\description{
Scales a block of data to unit variance
}
