% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PLS_variances_explained.R
\name{calculate_PLS_variances_explained}
\alias{calculate_PLS_variances_explained}
\title{Gets the variances explained from predictor LVs through direct effects, to
target LVs for an enire path model.}
\usage{
calculate_PLS_variances_explained(model, scaling = "numerical")
}
\arguments{
\item{model}{A path_model model estimated using the process_pls wrapper.}
}
\value{
a matrix, the same shape as the connection matrix, where the non zero
  elements indicate the explained variance from one node to another. The
  matrix is lower triangular.
}
\description{
Variance explained is calculated by looking at the ratio between the Sum of
Squares of the difference between Y and predicted Y through the direct
effects, and the Sum of Squares of Y. This ratio is subtracted from 1 to
yield the variance explained between 0 and 1. Note that negative values are
possible if more variance is introduced by prediction. This indicates that a
certain effect is not predictive, and one should consider to remove the
connection in the model definition.
}
