% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_SSE_for_matrices.R
\name{calculate_SSE_for_matrices}
\alias{calculate_SSE_for_matrices}
\title{Calculate the Sum of Squared Errors of two matrices}
\usage{
calculate_SSE_for_matrices(X1, X2)
}
\arguments{
\item{X1}{A matrix of m-by-n}

\item{X2}{A matrix of m-by-p}
}
\value{
The sum of squared errors (the difference) of the two matrices.
}
\description{
Calculates the Sum of Squared Errors of the two given matrices. If the
matrices do not have the same number of columns, the larger one is subsetted
from the first column onward to yield a matrix of equal dimensions. This
means that not all columns may be incorporated in the calculation.
}
\examples{
calculate_SSE_for_matrices(matrix(1, nrow=2, ncol=2), matrix(-1, nrow=2, ncol=2))
}
