% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arguments}
\alias{check_arguments}
\title{Checks the input arguments of the path_model() function}
\usage{
check_arguments(data, connection_matrix, variables_in_block, block_names,
  estimators)
}
\value{
A list of lists of errors and warnings respectively that indicate
  what is wrong with the arguments that are entered into the path_model function.
}
\description{
Checks each input argument in sequence, whilst suppressing errors, to allow for one-step checking of arguments.
}
