% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dag.R
\name{check_dag}
\alias{check_dag}
\title{Check Cyclicity in a Directed Acyclic Graph}
\usage{
check_dag(DAG, node_name, visited_nodes = list())
}
\arguments{
\item{DAG}{A matrix where the non zero elements what connections exist. The
rows indicate the node where the edge is going to, and the columns
indicates the node where the edge is coming from.}

\item{node_name}{A string indicating the identifier of the node from which
cyclicity is checked.}

\item{visited_nodes}{A list of strings indicating all the nodes that have
been visited in the recursive loop.}
}
\value{
A boolean indicating whether a cycle is detected (TRUE), or not
  (FALSE)
}
\description{
Checks whether a given connection matrix is acyclic. The function is
recursive and shrinks the search space until a solution is found.
}
\details{
This is an internal function called by verify_dag()
}
