% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_and_mask.R
\name{combine_and_mask}
\alias{combine_and_mask}
\title{Combines data from predictor and target nodes and constructs a mask for the covariance matrix accordingly}
\usage{
combine_and_mask(node, test_indices = NULL)
}
\arguments{
\item{node}{An object of the R6Class Node which is initialised and estimated at least once. The node needs to be connected to at least one target node for the Y matrices to exist.}

\item{test_indices}{a vector of indices indicating the row indices of the samples that belong to the test set. The training set is assumed to be the complementary set of data.}
}
\value{
list("X"=X,
             "Y"=Y,
             "covariance_mask"=covariance_mask,
             "cols_per_X_node"=cols_per_X_node,
             "cols_per_y_node"=cols_per_Y_node,
             "same_level_nodes"=same_level_nodes,
             "next_level_nodes"=next_level_nodes))

Or, when test_indices is supplied:

        list("X_train"=X_train,
             "X_test"=X_test,
             "Y_train"=Y_train,
             "Y_test"=Y_test,
             "covariance_mask"=covariance_mask,
             "cols_per_X_node"=cols_per_X_node,
             "cols_per_y_node"=cols_per_Y_node,
             "same_level_nodes"=same_level_nodes,
             "next_level_nodes"=next_level_nodes))

This is named list containing X, Y, the covariance mask, all nodes on the same level (predictors), all nodes on the next level (targets), and which columns of X and Y belong to which node originally. When test_indices is supplied, X and Y are split into X_train, X_test, Y_train, and Y_test
}
\description{
This function combines all the X_data from nodes on the same level as the entered node into a matrix X. The data is preprocessed with respect to a proper train and test split if test_indices is set.
All the previous_LVs from all target nodes are combined, although they are not processed further, into Y.
A covariance mask is also calculated, so that not all predictors are only predictive for targets that are connected according to the graph structure.
}
\details{
When test_indices are supplied, X and Y are split into a separate training and test set and are preprocessed accordingly.

combine_and_mask() was made for use with an iterative procedure, it constructs the targets based on the previously estimated LVs.
}
