% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\docType{data}
\name{ConvergenceLogger}
\alias{ConvergenceLogger}
\title{Logs the error per node per iteration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ConvergenceLogger
}
\description{
The logger show() method produces a plot of the iterations, on the X-axis, vs
the error of each iteration, on the Y-axis.
}
\keyword{datasets}
