% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\name{DurationLogger}
\alias{DurationLogger}
\title{Logs the duration of each iteration}
\description{
The logger show() method produces a plot of the iterations, on the X-axis, vs
the duration of each iteration, on the Y-axis.
}
\details{
When \code{report} is set to \code{TRUE} the duration is also reported each
iteration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DurationLogger$new()}}
\item \href{#method-log_status}{\code{DurationLogger$log_status()}}
\item \href{#method-show}{\code{DurationLogger$show()}}
\item \href{#method-clone}{\code{DurationLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationLogger$new(report = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_status"></a>}}
\if{latex}{\out{\hypertarget{method-log_status}{}}}
\subsection{Method \code{log_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationLogger$log_status(nodes, iteration)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationLogger$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DurationLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
