% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimator_and_initializer_string_to_node_class_type.R
\name{estimator_and_initializer_string_to_node_class_type}
\alias{estimator_and_initializer_string_to_node_class_type}
\title{Converts estimator and initializer string names to Node class type}
\usage{
estimator_and_initializer_string_to_node_class_type(estimator_name,
  initializer_name)
}
\arguments{
\item{estimator_name}{A string indicating the name of the estimator method.}

\item{initializer_name}{A string indicating the name of the initializer method.}
}
\value{
A R6Class Node, or a R6Class which inherits from Node.
}
\description{
This is an internal function used to convert each implemented estimator or
initializer name to its Node class type.
}
\details{
Supported options are: "normalpls" and "endpls" Other estimators are set to
the general R6Class Node. The conversion renders the name case-insensitive.
}
