% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_string_to_function.R
\name{estimator_string_to_function}
\alias{estimator_string_to_function}
\title{Converts estimator string names to function handle}
\usage{
estimator_string_to_function(estimator_name, parallelise = FALSE,
  n_cores = NULL)
}
\arguments{
\item{estimator_name}{A string indicating the name of the estimator method.}
}
\value{
A function handle to an estimator function which takes a node as an
  argument.
}
\description{
This is an internal function used to convert each implemented estimator name
to its corresponding function/closure.
}
\details{
Supported options are: "none". The conversion renders the name
case-insensitive.
}
