% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_functions.R
\name{full_estimator}
\alias{full_estimator}
\title{Estimate a node using its full data matrix}
\usage{
full_estimator(node)
}
\arguments{
\item{node}{An object of the R6Class Node which is initialised.}
}
\description{
Estimates the LVs of the given node by setting it to be equal to the full data matrix.
}
\details{
The function has no explicit return value, instead the R6Class node object is updated.
}
