% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer_functions.R
\name{full_initializer}
\alias{full_initializer}
\title{Initialise a node using its full data matrix}
\usage{
full_initializer(node)
}
\arguments{
\item{node}{An object of the R6Class Node.}
}
\description{
Initialises the LVs of the given node by setting it to be equal to the full
data matrix.
}
\details{
The function has no explicit return value, instead the R6Class node object is
updated.
}
