% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_node_connection_types.R
\name{get_all_node_connection_types}
\alias{get_all_node_connection_types}
\title{Construct a list of all node types in a connection matrix}
\usage{
get_all_node_connection_types(connection_matrix)
}
\arguments{
\item{connection_matrix}{A matrix where the non zero elements what
connections exist. The rows indicate the node where the edge is going to,
and the columns indicates the node where the edge is coming from.}
}
\value{
A list of strings indicating the node type for each node specified in
  the \code{connection_matrix}
}
\description{
Constructs a list of node types for all nodes in a given connection matrix.
}
