% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_estimator_list.R
\name{get_estimator_list}
\alias{get_estimator_list}
\title{Converts list of strings indicating node types to list of estimator function
handles}
\usage{
get_estimator_list(node_connection_types, start_node_estimator,
  middle_node_estimator, end_node_estimator, parallelise = FALSE,
  n_cores = NULL)
}
\arguments{
\item{node_connection_types}{A list of strings indicating the type of the node}

\item{start_node_estimator}{A handle to a function which takes a node as an
argument and is able to estimate a Start type node}

\item{middle_node_estimator}{A handle to a function which takes a node as an
argument and is able to estimate a Middle type node}

\item{end_node_estimator}{A handle to a function which takes a node as an
argument and is able to estimate an End type node}
}
\value{
A list of function handles to an estimator function which takes a
  node as an argument.
}
\description{
This is an internal function used to make a list of estimators for each type
of node.
}
