% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initialization_list.R
\name{get_initialization_list}
\alias{get_initialization_list}
\title{Converts list of strings indicating node types to list of initializer
function handles}
\usage{
get_initialization_list(node_types, start_node_initialization,
  middle_node_initialization, end_node_initialization)
}
\arguments{
\item{node_types}{A list of strings indicating the type of the node}

\item{start_node_estimator}{A handle to a function which takes a node as an
argument and is able to initialize a Start type node}

\item{middle_node_estimator}{A handle to a function which takes a node as an
argument and is able to initialize a Middle type node}

\item{end_node_estimator}{A handle to a function which takes a node as an
argument and is able to initialize an End type node}
}
\value{
A list of function handles to an initializers function which takes a
  node as an argument.
}
\description{
This is an internal function used to make a list of initializers for each
type of node.
}
