% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LVs.R
\name{get_LVs}
\alias{get_LVs}
\title{Calculates the latent variables using the estimators for each node in an
iterative manner.}
\usage{
get_LVs(nodes, max_iterations, loggers)
}
\arguments{
\item{nodes}{A listenv of objects of the R6Class Node which are initialised.
The nodes also need to be connected.}

\item{max_iterations}{An integer indicating the maximum number of iterations
before execution of LV estimation is halted when the convergence criterion
is not met beforehand.}

\item{loggers}{A listenv of R6Class based loggers and/or reporters. All
loggers or reports must implement a \code{log_status()} method. Implemented
loggers or reporters are ComponentLogger, IterationReporter,
DurationLogger, and ConvergenceLogger.}
}
\value{
A listenv of connected, initialized, and estimated nodes
}
\description{
This is an internal function used for the iterative estimation of the Latent
Variables.
}
\details{
A convergence threshold of a differnece of 0.0001 between the Sum of Squared
Errors of two subsequent iterations is used.
}
