% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_class_types.R
\name{get_node_class_types}
\alias{get_node_class_types}
\title{Converts list of strings indicating node types to list of Node class types}
\usage{
get_node_class_types(node_connection_types, start_node_estimator,
  middle_node_estimator, end_node_estimator, start_node_initializer,
  middle_node_initializer, end_node_initializer)
}
\arguments{
\item{node_connection_types}{A list of strings indicating the type of the node}

\item{start_node_estimator}{A handle to a function which takes a node as an
argument and is able to estimate a Start type node}

\item{middle_node_estimator}{A handle to a function which takes a node as an
argument and is able to estimate a Middle type node}

\item{end_node_estimator}{A handle to a function which takes a node as an
argument and is able to estimate an End type node}
}
\value{
A list of node class types
}
\description{
This is an internal function used to make a list of class types for each
connection type of node.
}
