% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_connection_type.R
\name{get_node_connection_type}
\alias{get_node_connection_type}
\title{Determines the type of a node}
\usage{
get_node_connection_type(connection_matrix, node_index)
}
\arguments{
\item{connection_matrix}{A matrix where the non zero elements what
connections exist. The rows indicate the node where the edge is going to,
and the columns indicates the node where the edge is coming from.}

\item{node_index}{a number indicating of which node the type is being
determined.}
}
\value{
A string indicating the type of node using "Start", "End", or
  "Middle"
}
\description{
Determines whether a node is a Start, Middle, or End type node.
}
