% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodes_by_level.R
\name{get_nodes_by_level}
\alias{get_nodes_by_level}
\title{Gets a listenv of listenvs of same level, and next level nodes}
\usage{
get_nodes_by_level(node)
}
\arguments{
\item{node}{An object of the R6Class Node which is initialised.}
}
\value{
A listenv with two listenvs of nodes on the same level, and nodes on
  the next level.
}
\description{
Constructs a listenv with length 2 which contains a listenv of all nodes on
the same level, and a listenv of all nodes on the next level.
}
