% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_previous_node_indices.R
\name{get_previous_node_indices}
\alias{get_previous_node_indices}
\title{Gets the indices of the previous nodes in a connection matrix given a node_index}
\usage{
get_previous_node_indices(connection_matrix, node_index)
}
\arguments{
\item{connection_matrix}{A matrix where the non zero elements what
connections exist. The rows indicate the node where the edge is going to,
and the columns indicates the node where the edge is coming from.}

\item{node_index}{a number indicating of which node the type is being
determined.}
}
\value{
A list of indices of all previous level nodes
}
\description{
Gets the integer indices of the previous nodes given a connection_matrix and a
starting index.
}
