% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_same_level_connected_nodes.R
\name{get_same_level_connected_nodes}
\alias{get_same_level_connected_nodes}
\title{Gets a listenv same level nodes}
\usage{
get_same_level_connected_nodes(node)
}
\arguments{
\item{node}{An object of the R6Class Node which is initialised.}
}
\value{
A listenv of nodes on the same level.
}
\description{
Constructs a listenv which contains  all nodes on
the same level
}
