% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_nodes.R
\name{get_unique_nodes}
\alias{get_unique_nodes}
\title{Create a listenv of all unique nodes given a listenv of nodes}
\usage{
get_unique_nodes(all_nodes)
}
\arguments{
\item{all_nodes}{A listenv of nodes}
}
\value{
A listenv of nodes. It contains a single reference to each unique
  node in all_nodes.
}
\description{
constructs a list of all unique nodes in the input, which may contain
duplicates.
}
