% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer_string_to_function.R
\name{initializer_string_to_function}
\alias{initializer_string_to_function}
\title{Converts initializer string names to function handle}
\usage{
initializer_string_to_function(initializer_name)
}
\arguments{
\item{initializer_name}{A string indicating the name of the initialiser method.}
}
\value{
A function handle to an initialiser function which takes a node as an
  argument.
}
\description{
This is an internal function used to convert each implemented initialiser
name to its corresponding function handle.
}
\details{
Supported options are: PCA and Full. The conversion renders the name
case-insensitive.
}
