% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\name{IterationReporter}
\alias{IterationReporter}
\title{Reports the number of each iteration.}
\description{
Reports the number of each iteration.

Reports the number of each iteration.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{IterationReporter$new()}}
\item \href{#method-log_status}{\code{IterationReporter$log_status()}}
\item \href{#method-clone}{\code{IterationReporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IterationReporter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_status"></a>}}
\if{latex}{\out{\hypertarget{method-log_status}{}}}
\subsection{Method \code{log_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IterationReporter$log_status(nodes, iteration)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IterationReporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
