% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_preprocessing_functions.R
\name{mean_center}
\alias{mean_center}
\title{Mean-centers each variable in a block of data}
\usage{
mean_center(input_data, settings = list(column_means = NULL))
}
\arguments{
\item{input_data}{A matrix containing data. The rows indicate the samples,
the columns indicate the variables.}

\item{settings}{A list with a one named elements "column_means". When
"column_means" is set, the supplied vector is used for mean substraction.
The vectors should have a length equal to the number of variables.}
}
\value{
A named list with "preprocessed_data": the mean-centered input_data
  matrix, and "settings", which is equal to the settings input parameter when
  set, and otherwise contains the calculated column_means.
}
\description{
Mean-centers each variable in a block of data
}
