% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{MSE}
\alias{MSE}
\title{Calculates the Mean Squared Error of two vectors or matrices}
\usage{
MSE(target, predicted)
}
\arguments{
\item{target}{A m-by-n matrix or length m vector of target values.}

\item{predicted}{A m-by-n matrix or length m vectorof predicted values.}
}
\value{
A Double indicating the Mean Squared Error
}
\description{
Calculates the Mean of the squared errors, or the difference between the two
input vectors or matrices.
}
\details{
When the number is higher than the number of variables in the internal X
matrix, the function might display unwanted behaviour.
}
