% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer_functions.R
\name{normal_SOPLS_initializer}
\alias{normal_SOPLS_initializer}
\title{n_LVs should be a list of n_LVs per block}
\usage{
normal_SOPLS_initializer(
  node,
  n_LVs_per_block = NULL,
  parallelise = FALSE,
  n_cores = NULL,
  error_function = SSE,
  LV_selection_method = "minimum_mean"
)
}
\description{
n_LVs should be a list of n_LVs per block
}
