% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_model.R
\name{path_model}
\alias{path_model}
\title{Performs Latent Variable estimation given data and path model specifications}
\usage{
path_model(data, connection_matrix, variables_in_block,
  block_names = NULL, estimators = NULL,
  start_node_estimator = "PLS", middle_node_estimator = "PLS",
  end_node_estimator = "PLS", initializers = NULL,
  start_node_initializer = "Full", middle_node_initializer = "Full",
  end_node_initializer = "Full", max_iterations = 100,
  loggers = NULL, unique_node_preprocessing = FALSE,
  global_preprocessors = list(),
  local_preprocessors = list(standardize))
}
\description{
Performs Latent Variable estimation given data and path model specifications
}
