% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_functions.R
\name{PCA_estimator}
\alias{PCA_estimator}
\title{Estimate a node using PCA}
\usage{
PCA_estimator(node, n_LVs = NULL)
}
\arguments{
\item{node}{An object of the R6Class Node which is initialised.}

\item{n_LVs}{An integer indicating the rank, or maximum number of LVs/PCs. If left as \code{NULL}, the rank is automatically estimated.}
}
\description{
Estimates the LVs of the given node using PCA estimation.
}
\details{
The number of LVs or the \code{rank} can be set explicitly, or it will be
automatically determined.

The function has no explicit return value, instead the R6Class node object is
updated.
}
