% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializer_functions.R
\name{PCA_initializer}
\alias{PCA_initializer}
\title{Initialises a node using PCA}
\usage{
PCA_initializer(node, rank = NULL)
}
\arguments{
\item{node}{An object of the R6Class Node.}

\item{rank}{An integer indicating the rank, or maximum number of LVs/PCs. If
left as \code{NULL}, the rank is equal to the number of variables in that
block.}
}
\description{
Initialises the LVs of the given node using PCA.
}
\details{
The number of LVs or the \code{rank} can be set explicitly, or it will be
automatically determined.

The function has no explicit return value, instead the R6Class node object is
updated.

The PCA initializer suffers from the sign indeterminacy. This is an issue
that will be adressed in future releases.
}
