% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_inner_model_variances}
\alias{plot_inner_model_variances}
\title{Plot the variances explained through the path}
\usage{
plot_inner_model_variances(model)
}
\description{
Plot the variances explained through the path
}
