% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_outer_effects}
\alias{plot_outer_effects}
\title{If combine is true: plot all outer model weights in a single plot
Default plots}
\usage{
plot_outer_effects(model, what_node = NULL, combine = TRUE,
  negative_values = "absolute")
}
\description{
If combine is true: plot all outer model weights in a single plot
Default plots
}
