% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_variances}
\alias{plot_variances}
\title{Plot the variances explained through the path}
\usage{
plot_variances(model, mode = "circle")
}
\description{
Plot the variances explained through the path
}
