% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator_functions.R
\name{PLS_estimator}
\alias{PLS_estimator}
\title{Estimate a node using PLS regression}
\usage{
PLS_estimator(node)
}
\arguments{
\item{node}{An object of the R6Class Node which is initialised and estimated
at least once. The node needs to be connected to at least one target node
for the Y matrices to exist.}
}
\description{
Estimates the LVs of the given node, and all nodes on the same level, using
PLS regression with all next level nodes as targets.
}
\details{
The cross product matrix used in SIMPLS regression is set to zero at the
connections between the target and predictor variables which are not defined
as being connected in the path model specification.

The function has no explicit return value, instead the R6Class node object,
and all same level connected nodes, are updated.
}
