% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_classes.R
\name{PLSPathNode}
\alias{PLSPathNode}
\title{PathNode is a template Node class for nodes which propagate effects, such as
the PLSNode. Path nodes should implement the get_paths_through_self_to_node method}
\description{
PathNode is a template Node class for nodes which propagate effects, such as
the PLSNode. Path nodes should implement the get_paths_through_self_to_node method

PathNode is a template Node class for nodes which propagate effects, such as
the PLSNode. Path nodes should implement the get_paths_through_self_to_node method
}
\section{Super classes}{
\code{\link[pathmodelr:DataNode]{pathmodelr::DataNode}} -> \code{\link[pathmodelr:LVNode]{pathmodelr::LVNode}} -> \code{PLSPathNode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_paths_to_self}{\code{PLSPathNode$get_paths_to_self()}}
\item \href{#method-get_node_path_effects}{\code{PLSPathNode$get_node_path_effects()}}
\item \href{#method-get_outgoing_path_to_node}{\code{PLSPathNode$get_outgoing_path_to_node()}}
\item \href{#method-get_paths_through_self_to_node}{\code{PLSPathNode$get_paths_through_self_to_node()}}
\item \href{#method-clone}{\code{PLSPathNode$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="DataNode" data-id="add_connected_nodes">}\href{../../pathmodelr/html/DataNode.html#method-add_connected_nodes}{\code{pathmodelr::DataNode$add_connected_nodes()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="DataNode" data-id="call_initializer">}\href{../../pathmodelr/html/DataNode.html#method-call_initializer}{\code{pathmodelr::DataNode$call_initializer()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="DataNode" data-id="preprocess_X">}\href{../../pathmodelr/html/DataNode.html#method-preprocess_X}{\code{pathmodelr::DataNode$preprocess_X()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="DataNode" data-id="preprocess_train_test">}\href{../../pathmodelr/html/DataNode.html#method-preprocess_train_test}{\code{pathmodelr::DataNode$preprocess_train_test()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="LVNode" data-id="add_estimate">}\href{../../pathmodelr/html/LVNode.html#method-add_estimate}{\code{pathmodelr::LVNode$add_estimate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="LVNode" data-id="estimate">}\href{../../pathmodelr/html/LVNode.html#method-estimate}{\code{pathmodelr::LVNode$estimate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pathmodelr" data-topic="LVNode" data-id="initialize">}\href{../../pathmodelr/html/LVNode.html#method-initialize}{\code{pathmodelr::LVNode$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_paths_to_self"></a>}}
\if{latex}{\out{\hypertarget{method-get_paths_to_self}{}}}
\subsection{Method \code{get_paths_to_self()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLSPathNode$get_paths_to_self()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_node_path_effects"></a>}}
\if{latex}{\out{\hypertarget{method-get_node_path_effects}{}}}
\subsection{Method \code{get_node_path_effects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLSPathNode$get_node_path_effects()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_outgoing_path_to_node"></a>}}
\if{latex}{\out{\hypertarget{method-get_outgoing_path_to_node}{}}}
\subsection{Method \code{get_outgoing_path_to_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLSPathNode$get_outgoing_path_to_node(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_paths_through_self_to_node"></a>}}
\if{latex}{\out{\hypertarget{method-get_paths_through_self_to_node}{}}}
\subsection{Method \code{get_paths_through_self_to_node()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLSPathNode$get_paths_through_self_to_node(node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLSPathNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
