% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{process_PLS}
\alias{process_PLS}
\title{Estimate a process PLS model.}
\usage{
process_PLS(data, connection_matrix, variables_in_block,
  block_names = NULL, loggers = listenv(IterationReporter$new(),
  DurationLogger$new(report = TRUE)), max_iterations = 100,
  global_preprocessors = list(),
  local_preprocessors = list(standardize))
}
\arguments{
\item{data}{A dataframe or matrix containing the data belonging to the
process that is being modelled. The rows contain the samples, the columns
contain the variables.}

\item{connection_matrix}{A lower triangular matrix where the non zero
elements what connections exist. The rows indicate the node where the edge
is going to, and the columns indicates the node where the edge is coming
from.}

\item{variables_in_block}{A list of lists indicating the variables
corresponding to each block. The ordering of the upper list should be the
same as that of the \code{connection_matrix}. The inner list values can be
integers indicating the column number, or the name when column names are
set in \code{data}.}

\item{block_names}{An ordered list containing the names that should be
assigned to each block. The ordering of the list should be the same as that
of the \code{connection_matrix}. When the argument is not set, names are
extracted from \code{data}, or set to dummy names based on ordering when
\code{data} contains no names.}

\item{loggers}{A listenv of R6Class based logger and/or reporter objects. All
loggers or reports must implement a \code{log_status()} method. Implemented
loggers or reporters are ComponentLogger, IterationReporter, DurationLogger,
and ConvergenceLogger.}

\item{max_iterations}{An integer indicating the maximum number of iterations
before execution of LV estimation is halted when the convergence criterion
is not met beforehand.}

\item{global_preprocessors}{A list of preprocessing functions when
unique_node_preprocessing is \code{FALSE}, a list of lists of preprocessing
functions when unique_node_preprocessing is \code{TRUE}. The ordering of
the outer list should be the same as that of the \code{connection_matrix}.
The order of the inner list is only important when the function results
differ when their order of application is changed. The funtions will be
applied from beginning to end. The supplied function must be invariant to
subsampling. User-implemented functions must take a Matrix as input, and
return the preprocessed matrix.}

\item{local_preprocessors}{A list of preprocessing functions when
unique_node_preprocessing is \code{FALSE}, a list of lists of preprocessing
functions when unique_node_preprocessing is \code{TRUE}. The ordering of
the outer list should be the same as that of the \code{connection_matrix}.
The order of the inner list is only important when the function results
differ when their order of application is changed. The funtions will be
applied from beginning to end. The supplied functionis assumed to be
influenced by subsampling, and therefore can be applied on subsets when
(cross-)validating. User-implemented functions must take a Matrix as input,
and return the preprocessed matrix. Implemented functions are: (block_scale,
standardize, and mean_center)}
}
\description{
Estimate a process PLS model.
}
