% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_preprocessing_functions.R
\name{standardize}
\alias{standardize}
\title{Mean-centers each variable in a block of data and then scale each variable to
unit variance}
\usage{
standardize(input_data, settings = list(column_means = NULL, column_stds
  = NULL))
}
\arguments{
\item{input_data}{A matrix containing data. The rows indicate the samples,
the columns indicate the variables.}

\item{settings}{A list with a two named elements "column_stds" and
"column_means". When "column_stds" is set, the supplied vector is used for
scaling, otherwise it is calculated. When "column_means" is set, the
supplied vector is used for mean substraction. Both vectors should have a
length equal to the number of variables. All named elements can be set, and
optionally left out, seperately.}
}
\value{
A named list with "preprocessed_data": the scaled input_data matrix,
  and "settings", which is equal to the settings input parameter when set,
  and otherwise contains the calculated and used column_stds and
  column_means.
}
\description{
Mean-centers each variable in a block of data and then scale each variable to
unit variance
}
