% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_dag.R
\name{verify_dag}
\alias{verify_dag}
\title{Verify a Directed Acyclic Graph}
\usage{
verify_dag(DAG)
}
\arguments{
\item{DAG}{A lower triangular matrix where the non zero elements what connections exist. The
rows indicate the node where the edge is going to, and the columns
indicates the node where the edge is coming from.}
}
\value{
A list of lists of errors and warnings respectively that indicate
  what conditions of the DAG requirements are not met.
}
\description{
Verifies whether a given connection matrix is a Directed Acyclic Graph.
verify_dag wraps check_dag for a recursive Depth First Search for acyclicity.
}
\examples{
verify_dag(t(matrix(c(0,0,0,0,1,0,0,0,1,0,0,0,0,1,1,0),nrow=4, ncol=4)))
}
