% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\docType{data}
\name{ComponentLogger}
\alias{ComponentLogger}
\title{Logs the number of components per node per iteration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ComponentLogger
}
\description{
Logs the number of components per node per iteration
}
\keyword{datasets}
