% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\name{ConvergenceLogger}
\alias{ConvergenceLogger}
\title{Logs the error per node per iteration}
\description{
The logger show() method produces a plot of the iterations, on the X-axis, vs
the error of each iteration, on the Y-axis.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConvergenceLogger$new()}}
\item \href{#method-log_status}{\code{ConvergenceLogger$log_status()}}
\item \href{#method-show}{\code{ConvergenceLogger$show()}}
\item \href{#method-clone}{\code{ConvergenceLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvergenceLogger$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_status"></a>}}
\if{latex}{\out{\hypertarget{method-log_status}{}}}
\subsection{Method \code{log_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvergenceLogger$log_status(nodes, iteration)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvergenceLogger$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConvergenceLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
