% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\docType{data}
\name{DurationLogger}
\alias{DurationLogger}
\title{Logs the duration of each iteration}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DurationLogger
}
\description{
The logger show() method produces a plot of the iterations, on the X-axis, vs
the duration of each iteration, on the Y-axis.
}
\details{
When \code{report} is set to \code{TRUE} the duration is also reported each
iteration.
}
\keyword{datasets}
