% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers_reporters.R
\docType{data}
\name{IterationReporter}
\alias{IterationReporter}
\title{Reports the number of each iteration.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IterationReporter
}
\description{
Reports the number of each iteration.
}
\keyword{datasets}
